+++ components/select_menu/select_menu.test.js
@@ -1,5 +1,4 @@
 import { mount } from '@vue/test-utils';
-import { itBehavesLikeEmitsExpectedEvent } from '../../tests/shared_examples/events';
 import {
   itBehavesLikePassesCustomPropValidation,
   itBehavesLikeFailsCustomPropValidation,
@@ -8,10 +7,6 @@ import {
   initializeSpy,
   cleanSpy,
 } from '../../tests/shared_examples/validation';
-import {
-  itBehavesLikeAppliesClassToChild,
-  itBehavesLikeAppliesChildProp,
-} from '../../tests/shared_examples/extendability';
 import {
   LABEL_SIZE_MODIFIERS,
   DESCRIPTION_SIZE_MODIFIERS,
@@ -326,10 +321,10 @@ describe('DtSelectMenu Tests', () => {
       });
 
       it('should emit input event', () => {
-        itBehavesLikeEmitsExpectedEvent(wrapper, 'input', selectedValue.toString());
+        expect(wrapper.emitted('input')[0][0]).toBe(selectedValue.toString());
       });
       it('should emit change event', () => {
-        itBehavesLikeEmitsExpectedEvent(wrapper, 'change', selectedValue.toString());
+        expect(wrapper.emitted('change')[0][0]).toBe(selectedValue.toString());
       });
     });
   });
@@ -425,19 +420,6 @@ describe('DtSelectMenu Tests', () => {
       element = wrapper.find(selector);
     };
 
-    // Shared Examples
-    const itBehavesLikeAppliesClassToChildLocal = () => {
-      it('should apply custom class to child', () => {
-        itBehavesLikeAppliesClassToChild(wrapper, '.my-custom-class', element);
-      });
-    };
-
-    const itBehavesLikeAppliesChildPropLocal = () => {
-      it('should have provided child prop', () => {
-        itBehavesLikeAppliesChildProp(element, propName, propValue);
-      });
-    };
-
     // Test Setup
     beforeAll(() => {
       childProps[propName] = propValue;
@@ -450,21 +432,27 @@ describe('DtSelectMenu Tests', () => {
       beforeEach(
         () => { _setupChildClassTest('labelClass', '[data-qa="dt-select-label"]'); },
       );
-      itBehavesLikeAppliesClassToChildLocal();
+      it('should apply custom class to child', () => {
+        expect(wrapper.find('.my-custom-class').html()).toBe(element.html());
+      });
     });
 
     describe('When a description class is provided', () => {
       beforeEach(
         () => { _setupChildClassTest('descriptionClass', '[data-qa="dt-select-description"]'); },
       );
-      itBehavesLikeAppliesClassToChildLocal();
+      it('should apply custom class to child', () => {
+        expect(wrapper.find('.my-custom-class').html()).toBe(element.html());
+      });
     });
 
     describe('When a select class is provided', () => {
       beforeEach(
         () => { _setupChildClassTest('selectClass', '[data-qa="dt-select-wrapper"]'); },
       );
-      itBehavesLikeAppliesClassToChildLocal();
+      it('should apply custom class to child', () => {
+        expect(wrapper.find('.my-custom-class').html()).toBe(element.html());
+      });
     });
 
     describe('When an option class is provided', () => {
@@ -489,14 +477,19 @@ describe('DtSelectMenu Tests', () => {
       beforeEach(
         () => { _setupChildPropsTest('labelChildProps', '[data-qa="dt-select-label"]'); },
       );
-      itBehavesLikeAppliesChildPropLocal();
+
+      it('should have provided child prop', () => {
+        expect(element.attributes(propName)).toBe(propValue);
+      });
     });
 
     describe('When description child props are provided', () => {
       beforeEach(
         () => { _setupChildPropsTest('descriptionChildProps', '[data-qa="dt-select-description"]'); },
       );
-      itBehavesLikeAppliesChildPropLocal();
+      it('should have provided child prop', () => {
+        expect(element.attributes(propName)).toBe(propValue);
+      });
     });
 
     describe('When option child props are provided', () => {
@@ -512,7 +505,7 @@ describe('DtSelectMenu Tests', () => {
 
       it('should apply child props to each option', () => {
         options.forEach(option => {
-          itBehavesLikeAppliesChildProp(option, propName, propValue);
+          expect(option.attributes(propName)).toBe(propValue);
         });
       });
     });
@@ -525,7 +518,9 @@ describe('DtSelectMenu Tests', () => {
         element = select;
       });
 
-      itBehavesLikeAppliesChildPropLocal();
+      it('attr should be set on element', () => {
+        expect(element.attributes(propName)).toBe(propValue);
+      });
     });
   });
 });